don't make other services worry about gcs storage
